rename getComponentActions to getActions
